+++ backend/ContainerApp/Engine/Models/Chat/ChatAiServiceResponse.cs
@@ -1,5 +1,5 @@
-﻿using System.Text.Json;
-using Engine.Services.Clients.AccessorClient.Models;
+﻿using Engine.Services.Clients.AccessorClient.Models;
+using Microsoft.SemanticKernel.ChatCompletion;
 
 namespace Engine.Models.Chat;
 
@@ -8,9 +8,8 @@ public record ChatAiServiceResponse
     public required string RequestId { get; init; }
     public required Guid ThreadId { get; init; }
     public ChatMessage? Answer { get; set; }
-    public string Name { get; set; } = string.Empty;
 
-    public JsonElement UpdatedHistory { get; set; }
+    public ChatHistory UpdatedHistory { get; set; } = new ChatHistory();
     public ChatAnswerStatus Status { get; set; } = ChatAnswerStatus.Ok;
     public string? Error { get; set; }
 }
