+++ backend/ContainerApp/UnitTests/EngineUnitTests/Endpoints/EngineQueueHandlerTests.cs
@@ -1,6 +1,7 @@
 ﻿using DotQueue;
 using Engine.Constants;
 using Engine.Endpoints;
+using Engine.Helpers;
 using Engine.Models;
 using Engine.Models.Chat;
 using Engine.Models.QueueMessages;
@@ -9,6 +10,7 @@ using Engine.Services.Clients.AccessorClient;
 using Engine.Services.Clients.AccessorClient.Models;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
+using Microsoft.SemanticKernel.ChatCompletion;
 using Moq;
 using System.Text.Json;
 
@@ -34,6 +36,7 @@ public class EngineQueueHandlerTests
         Mock<IChatAiService> ai,
         Mock<IAiReplyPublisher> pub,
         Mock<IAccessorClient> accessorClient,
+        Mock<IChatTitleService> titleService,
         Mock<ILogger<EngineQueueHandler>> log,
         EngineQueueHandler sut
     ) CreateSut()
@@ -42,16 +45,18 @@ public class EngineQueueHandlerTests
         var ai = new Mock<IChatAiService>(MockBehavior.Strict);
         var pub = new Mock<IAiReplyPublisher>(MockBehavior.Strict);
         var accessorClient = new Mock<IAccessorClient>(MockBehavior.Strict);
+        var titleService = new Mock<IChatTitleService>(MockBehavior.Strict);
         var log = new Mock<ILogger<EngineQueueHandler>>();
-        var sut = new EngineQueueHandler(engine.Object, ai.Object, pub.Object, accessorClient.Object, log.Object);
-        return (engine, ai, pub, accessorClient, log, sut);
+
+        var sut = new EngineQueueHandler(engine.Object, ai.Object, pub.Object, accessorClient.Object, titleService.Object, log.Object);
+        return (engine, ai, pub, accessorClient, titleService, log, sut);
     }
 
     [Fact]
     public async Task HandleAsync_CreateTask_Processes_TaskModel()
     {
         // Arrange
-        var (engine, ai, pub, accessorClient, log, sut) = CreateSut();
+        var (engine, ai, pub, accessorClient, titleSrvice, log, sut) = CreateSut();
 
         var task = new TaskModel
         {
@@ -82,7 +87,7 @@ public class EngineQueueHandlerTests
     [Fact]
     public async Task HandleAsync_CreateTask_InvalidPayload_DoesNotCall_Engine()
     {
-        var (engine, ai, pub, accessorClient, log, sut) = CreateSut();
+        var (engine, ai, pub, accessorClient, titleSrvice, log,  sut) = CreateSut();
 
         // payload = "null"
         var msg = new Message
@@ -105,7 +110,7 @@ public class EngineQueueHandlerTests
     public async Task HandleAsync_ProcessingQuestionAi_HappyPath_Calls_Ai_And_Publishes()
     {
         // Arrange
-        var (engine, ai, pub, accessorClient, log, sut) = CreateSut();
+        var (engine, ai, pub, accessorClient, titleSrvice, log, sut) = CreateSut();
 
         var requestId = Guid.NewGuid().ToString();
         var threadId = Guid.NewGuid();
@@ -136,27 +141,24 @@ public class EngineQueueHandlerTests
         accessorClient.Setup(a => a.GetHistorySnapshotAsync(threadId, userId, It.IsAny<CancellationToken>()))
                 .ReturnsAsync(snapshotFromAccessor);
 
+        var history = new ChatHistory();
+        history.AddSystemMessage("You are a helpful assistant.");
+        history.AddUserMessageNow(userMsg);
         var expectedAiReq = new ChatAiServiseRequest
         {
             RequestId = requestId,
             ThreadId = threadId,
-            UserMessage = userMsg,
             ChatType = ChatType.Default,
             UserId = userId,
             SentAt = engineReq.SentAt,
             TtlSeconds = 120,
-            History = snapshotFromAccessor.History
+            History = history
         };
 
-        var updatedHistory = Je("""
-        {
-          "messages":[
-            {"role":"system","content":"..."},
-            {"role":"user","content":"hello"},
-            {"role":"assistant","content":"world"}
-          ]
-        }
-        """);
+
+        var updatedHistory = history;
+
+        updatedHistory.AddAssistantMessage("World");
 
         var answer = new ChatMessage
         {
@@ -167,41 +169,29 @@ public class EngineQueueHandlerTests
             Content = textAnswer
         };
 
+        history.AddAssistantMessage(textAnswer);
+
         var aiResp = new ChatAiServiceResponse
         {
             RequestId = requestId,
             ThreadId = threadId,
             Status = ChatAnswerStatus.Ok,
             Answer = answer,
-            Name = chatName,
             UpdatedHistory = updatedHistory
         };
 
         ai.Setup(a => a.ChatHandlerAsync(It.Is<ChatAiServiseRequest>(r =>
                         r.RequestId == expectedAiReq.RequestId &&
                         r.ThreadId == expectedAiReq.ThreadId &&
-                        r.UserMessage == expectedAiReq.UserMessage &&
                         r.UserId == expectedAiReq.UserId &&
                         r.ChatType == expectedAiReq.ChatType &&
-                        r.History.GetRawText() == expectedAiReq.History.GetRawText()
+                        r.History[0].Content == expectedAiReq.History[0].Content
+                        &&
+                        r.History[1].Content == expectedAiReq.History[1].Content
+                        &&
+                        r.History[2].Content == expectedAiReq.History[2].Content
                     ), It.IsAny<CancellationToken>()))
           .ReturnsAsync(aiResp);
-
-        accessorClient.Setup(a => a.UpsertHistorySnapshotAsync(
-                        It.Is<UpsertHistoryRequest>(u =>
-                            u.ThreadId == threadId &&
-                            u.UserId == userId &&
-                            u.ChatType == "default" &&
-                            u.History.GetRawText() == updatedHistory.GetRawText()
-                        ),
-                        It.IsAny<CancellationToken>()))
-                .ReturnsAsync(new HistorySnapshotDto
-                {
-                    ThreadId = threadId,
-                    UserId = userId,
-                    ChatType = "default",
-                    History = updatedHistory
-                });
 
         var engineResponse = new EngineChatResponse
         {
@@ -240,7 +230,7 @@ public class EngineQueueHandlerTests
     [Fact(Skip = "Todo: do after refactoring ai Chat for queue")]
     public async Task HandleAsync_ProcessingQuestionAi_MissingThreadId_ThrowsNonRetryable_AndSkipsWork()
     {
-        var (engine, ai, pub, accessorClient, log, sut) = CreateSut();
+        var (engine, ai, pub, accessorClient, titleSrvice, log, sut) = CreateSut();
 
         var requestId = Guid.NewGuid().ToString();
         var userId = Guid.NewGuid();
@@ -284,11 +274,15 @@ public class EngineQueueHandlerTests
     public async Task HandleAsync_ProcessingQuestionAi_AiThrows_WrappedInRetryable()
     {
         // Arrange
-        var (engine, ai, pub, accessor, log, sut) = CreateSut();
+        var (engine, ai, pub, accessor, log, titleSrvice, sut) = CreateSut();
 
         var requestId = Guid.NewGuid().ToString();
         var threadId = Guid.NewGuid();
         var userId = Guid.NewGuid();
+
+        var history = new ChatHistory();
+        history.AddSystemMessage("You are a helpful assistant.");
+        history.AddUserMessageNow("boom");
 
         var engineReq = new EngineChatRequest
         {
@@ -312,21 +306,32 @@ public class EngineQueueHandlerTests
         accessor.Setup(a => a.GetHistorySnapshotAsync(threadId, userId, It.IsAny<CancellationToken>()))
                 .ReturnsAsync(snapshotFromAccessor);
 
+        accessor
+            .Setup(a => a.UpsertHistorySnapshotAsync(
+                It.IsAny<UpsertHistoryRequest>(),
+                It.IsAny<CancellationToken>()))
+            .ReturnsAsync(new HistorySnapshotDto
+            {
+                ThreadId = threadId,
+                UserId = userId,
+                ChatType = "default",
+                History = EmptyHistory()
+            });
+
         var expectedAiReq = new ChatAiServiseRequest
         {
             RequestId = requestId,
             ThreadId = threadId,
-            UserMessage = "boom",
             ChatType = ChatType.Default,
             UserId = userId,
             SentAt = engineReq.SentAt,
             TtlSeconds = 120,
-            History = snapshotFromAccessor.History
+            History = history
         };
 
         ai.Setup(a => a.ChatHandlerAsync(It.Is<ChatAiServiseRequest>(r =>
                         r.ThreadId == threadId &&
-                        r.UserMessage == "boom"
+                        r.History[1].Content == "boom"
                     ), It.IsAny<CancellationToken>()))
           .ThrowsAsync(new InvalidOperationException("AI failed"));
 
@@ -346,7 +351,6 @@ public class EngineQueueHandlerTests
         ex.InnerException!.Message.Should().Contain("AI failed");
 
         accessor.Verify(a => a.GetHistorySnapshotAsync(threadId, userId, It.IsAny<CancellationToken>()), Times.Once);
-        accessor.Verify(a => a.UpsertHistorySnapshotAsync(It.IsAny<UpsertHistoryRequest>(), It.IsAny<CancellationToken>()), Times.Never);
         pub.VerifyNoOtherCalls();
         engine.VerifyNoOtherCalls();
     }
@@ -354,7 +358,7 @@ public class EngineQueueHandlerTests
     [Fact]
     public async Task HandleAsync_UnknownAction_ThrowsNonRetryable_AndSkipsWork()
     {
-        var (engine, ai, pub, accessorClient, log, sut) = CreateSut();
+        var (engine, ai, pub, accessorClient, titleSrvice, log, sut) = CreateSut();
 
         var msg = new Message
         {
